Moves workflow to its own directory; adjusts trigger readme
This should make the integration repository work better with the website;
it expects the workflow to be in a distinct file (so it can be added to
a user's account from a URL).

This also splits the instructions so the triggers/ subdir contains
instructions only about the trigger, with a link to the complete workflow.